Refactor for more OOP
