+++ src/main/java/Ui/JohnUi.java
@@ -6,4 +6,8 @@ public class JohnUi {
     public void printLine() {
         System.out.println(LINE);
     }
+
+    public void showLoadingError() {
+        System.out.println("Error loading tasks from file. Initialising an empty task list.");
+    }
 }
